% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powercalc.R
\name{AR1}
\alias{AR1}
\title{Correlation matrix of an autoregressive model of order 1}
\usage{
AR1(p, rho)
}
\arguments{
\item{p}{dimension}

\item{rho}{correlation}
}
\value{
(p x p) matrix
}
\description{
An order-1 autoregressive correlation matrix is set up which is
  used for the examples on power and sample size calculation.
}
\examples{
  AR1(10, 0.2)
}
