% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powercalc.R
\name{coeff.beta.k}
\alias{coeff.beta.k}
\title{Ratio of expected value to variance of estimator}
\usage{
coeff.beta.k(k, beta.true, lambda, eigendec, n, weights = 1)
}
\arguments{
\item{k}{index of selected regression coefficient}

\item{beta.true}{(LEN p) vector of regression coefficients}

\item{lambda}{shrinkage parameter}

\item{eigendec}{eigenvalue decomposition of (p x p) correlation matrix
\code{R}}

\item{n}{sample size}

\item{weights}{vector (LEN p) of SNP-specific weights or scalar if weights
are equal for all SNPs; default value 1}
}
\value{
ratio
}
\description{
The ratio of expected value to standard deviation is calculated
  for the estimator of a selected regression coefficient.
}
\examples{
  ### correlation matrix (should depend on sire haplotypes)
  R <- AR1(100, rho = 0.1)
  eigendec <- eigen(R)
  pos <- c(14, 75)
  b <- rep(0, nrow(R)); b[pos] <- 0.35
  coeff.beta.k(14, b, 1200, eigendec, 100)
}
