% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ExpectMat}
\alias{ExpectMat}
\title{ExpectMat}
\usage{
ExpectMat(inMat)
}
\arguments{
\item{inMat}{[MATRIX] The paternal genotype matrix}
}
\value{
\describe{
  \item{\code{ExP.Fa}}{(N x p) matrix of expected values}
}
}
\description{
Expected value of paternally inherited allele
}
\examples{
 data(testdata)
 G <- Haplo2Geno(H.sire)
 E <- ExpectMat(G)
}
