% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.sire.R
\docType{data}
\name{H.sire}
\alias{H.sire}
\title{testdata: sire haplotypes}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 300 columns.
}
\usage{
H.sire
}
\description{
(2N x p) matrix of sire haplotypes for all chromosomes (2 lines
  per sire); unknown alleles are marked as 9
}
\keyword{datasets}
