% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Haplo2Geno}
\alias{Haplo2Geno}
\title{Conversion of haplotypes into genotypes}
\usage{
Haplo2Geno(inpMat)
}
\arguments{
\item{inpMat}{[MATRIX] haplotype matrix (2 lines per individual)}
}
\value{
\describe{
  \item{\code{outMa}}{(N x p) genotype matrix}
}
}
\description{
Haplotypes are converted into into genotypes without checking
 for missing values.
}
\examples{
 data(testdata)
 G <- Haplo2Geno(H.sire)
}
