% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matLD.R
\docType{data}
\name{matLD}
\alias{matLD}
\title{testdata: maternal linkage disequilibrium}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 300 rows and 300 columns.
}
\usage{
matLD
}
\description{
(p x p) matrix of maternal LD between pairs of p markers;
  matrix is block diagonal in case of multiple chromosomes
}
\keyword{datasets}
