% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.chr.R
\docType{data}
\name{pos.chr}
\alias{pos.chr}
\title{testdata: genetic map positions}
\format{
An object of class \code{list} of length 1.
}
\usage{
pos.chr
}
\description{
list of vectors of genetic map positions per chromosome
}
\keyword{datasets}
