% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmat.R
\name{CovMat}
\alias{CovMat}
\title{Calculation of covariance or correlation matrix}
\usage{
CovMat(linkDam, haploSire, nfam, pos_chr, corr = T)
}
\arguments{
\item{linkDam}{(p x p) matrix of maternal LD between pairs of p markers;
matrix is block diagonal in case of multiple chromosomes}

\item{haploSire}{(2N x p) matrix of sires haplotypes for all chromosomes
(2 lines per sire)}

\item{nfam}{vector (LEN N) containing number of progeny per sire or
scalar value in case of equal family size}

\item{pos_chr}{list (LEN number of chromosomes) of vectors (LEN number of
markers) of genetic positions in Morgan per chromosome}

\item{corr}{logical; \code{TRUE} (default) if output is correlation matrix or
\code{FALSE} if output is covariance matrix}
}
\value{
list (LEN 2) of matrix (DIM \eqn{p1} x \eqn{p1}) and vector
  (LEN \eqn{p1}) with \eqn{p1 \le p}
\describe{
  \item{\code{K}}{covariance matrix OR}
  \item{\code{R}}{correlation matrix}
  \item{\code{valid.snps}}{vector of SNP indices considered for covariance/
  correlation matrix}
}
}
\description{
The theoretical covariance between pairs of markers is calculated
  from either paternal haplotypes and maternal linkage disequilibrium (LD) or
  vise versa. A genetic map is required. The implementation relies on
  paternal half-sib families and biallelic markers such as single nucleotide
  polymorphisms (SNP).
}
\note{
If maternal half-sib families are used, the roles of sire/dam are swapped.
  Multiple families can be considered.

  Family size is used for weighting covariance terms in case of multiple
  half-sib families. It only matters if number of progeny differs.

{
  If you have maternal haplotypes (H.mothers; same format as H.sire)
  instead of maternal LD (matLD) then LD can be estimated from counting
  haplotype frequencies as:

  \code{matLD <- LDdam(inMat = H.mother, pos.chr)}

  If multiple chromosomes are considered, then, for instance:

  \code{pos.chr <- list(pos.snp.chr1, pos.snp.chr2, pos.snp.chr3)}
}
}
\examples{
  ### 1: INPUT DATA
  data(testdata)
  ### 2: COVARIANCE/CORRELATION MATRIX
  corrmat <- CovMat(matLD, H.sire, 100, pos.chr, corr = TRUE)
  ### 3: TAGSNPS FROM CORRELATION MATRIX
  bin <- tagSNP(corrmat$R)
  bin <- tagSNP(corrmat$R, 0.5)
}
\references{
Wittenburg, Bonk, Doschoris, Reyer (2019) "Design of Experiments
  for Fine-Mapping Quantitative Trait Loci in Livestock Populations"
  \url{https://doi.org/10.1101/2019.12.17.879106}
}
